[ClangImporter] Move code to ImportName.cpp
Refactors much of the import name determination from ClangImporter.cpp
into ImportName.cpp.

This is not a perfect separation, but it's a step in the right
direction. Long term, we'd like to get as much functionality off of
the ClangImporter::Implementation class and onto more specialized and
seperable (and stateless!) components.
